requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
PyYAML>=5.4.1
joblib>=1.0.1
jupyter>=1.0.0
notebook>=6.4.0
python-dotenv>=0.19.0
tqdm>=4.61.2
pytest>=6.2.5
black>=21.6b0
flake8>=3.9.2
isort>=5.9.2 